Tests/LibWeb/Text/expected/css/fill-stroke-fallback-relative-lengths.txt

fill: url("#missing-fill") rgb(0, 255, 0)
stroke: url("#missing-stroke") rgb(0, 255, 0)
